==> Credit Default Prediction/2_Code_Jupyter_NoteBooks/02_DataPreprocessing_Transformations_&_FeatureSelection.ipynb <==
### NoteBook Contents ( DataPreprocessing Transformations & FeatureSelection):
        > 1. Seperating Numeric and Categorical features, assigning categories to categories not in the dataset description 
        > 2. Splitting the Data into Training and Test Set
        > 3. Missing Values Imputation
        > 4. Data standardization
        > 5. Saving training and test data

#### Importing Essential Data handling libraries

import pandas as pd
import numpy as np

import matplotlib as mpl
import matplotlib.pyplot as plt
import seaborn as sns

import warnings
warnings.filterwarnings("ignore")

%matplotlib inline

import time
from math import floor
from time import perf_counter
from tqdm import tqdm

pd.set_option('display.max_columns', 100)
pd.set_option('display.float_format', lambda x: '%.3f' % x)

from IPython.core.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import OneHotEncoder
from sklearn.impute import KNNImputer
from sklearn.preprocessing import QuantileTransformer
from sklearn.feature_selection import VarianceThreshold
from sklearn.preprocessing import StandardScaler

#### Loading Data

df_master = pd.read_csv("..\\0_Modeling_Data\\data.csv") ## looking for dataset ,up one directory

df_raw = df_master.copy()

## Dropping "ID" field since its a unique identifier 
df_raw.drop("ID", axis = 1, inplace = True)

df_raw.shape 

df_raw.head()

### 1. Seperating Numeric and Categorical features : 

target_variable = "default payment next month"

## Assumption  : if the nuber of unique values in the feature are < 0.05 % of the total observations, then treat this as categorical feature

categorical_threshold = df_raw.shape[0]*0.05/100
print(categorical_threshold)

categorical_variables = df_raw.drop(target_variable, axis = 1).apply(lambda x: len(np.unique(x.value_counts()))).index[(df_raw.drop(target_variable, axis = 1).apply(lambda x: len(np.unique(x.value_counts()))) < categorical_threshold)]
categorical_variables = list(categorical_variables)
print(categorical_variables)

### predictors
numeric_variables = df_raw.drop([target_variable] + categorical_variables,axis = 1).columns.to_list()
print(numeric_variables)

print(len(categorical_variables))
print(len(numeric_variables))

##### Data Mapping in categorical features
for feat in categorical_variables:
    print(feat)
    print(df_raw[feat].value_counts())
    print("\n")

### Mapping unknown values to respective categories (as per data analysis)
df_raw["EDUCATION"] = df_raw["EDUCATION"].map({0:6, 1:1,2:2,3:3,4:4,5:5,6:6})
df_raw["MARRIAGE"] = df_raw["MARRIAGE"].map({0:3, 1:1,2:2,3:3})
df_raw["PAY_0"] = df_raw["PAY_0"].map({-2:0,-1:0,0:0,1:1,2:2,3:3,4:4,5:5,6:6,7:7,8:8,9:9})
df_raw["PAY_2"] = df_raw["PAY_2"].map({-2:0,-1:0,0:0,1:1,2:2,3:3,4:4,5:5,6:6,7:7,8:8,9:9})
df_raw["PAY_3"] = df_raw["PAY_3"].map({-2:0,-1:0,0:0,1:1,2:2,3:3,4:4,5:5,6:6,7:7,8:8,9:9})
df_raw["PAY_4"] = df_raw["PAY_4"].map({-2:0,-1:0,0:0,1:1,2:2,3:3,4:4,5:5,6:6,7:7,8:8,9:9})
df_raw["PAY_5"] = df_raw["PAY_5"].map({-2:0,-1:0,0:0,1:1,2:2,3:3,4:4,5:5,6:6,7:7,8:8,9:9})
df_raw["PAY_6"] = df_raw["PAY_6"].map({-2:0,-1:0,0:0,1:1,2:2,3:3,4:4,5:5,6:6,7:7,8:8,9:9})

categorical_variables

numeric_variables

### 2. Splitting the Data into Training and Test Set

X_train, X_test, y_train, y_test = train_test_split(df_raw.drop(target_variable, axis = 1), 
                                                   df_raw[target_variable], 
                                                    test_size = 0.20,
                                                    random_state = 8,
                                                   shuffle = True)

X_train.reset_index(inplace = True, drop  = True)
X_test.reset_index(inplace = True, drop  = True)

print(X_train.shape)
print(X_test.shape)

y_train.value_counts(normalize=True)

y_test.value_counts(normalize=True)

y_train = y_train.values
y_test = y_test.values

X_train.dtypes

##### Missing value imputation process
    > we have all the data in numeric terms, Impute missing values using Imputer (e.g. KNN)

### 3. Missing Values Imputation

missing_value_imputer = KNNImputer(n_neighbors = 11, weights="uniform")

##### fit imputer on trainig data

missing_value_imputer.fit(X_train)

##### Impute missing values in train and test data accordingly

t0 = perf_counter()

X_train = pd.DataFrame(missing_value_imputer.transform(X_train), columns = X_train.columns)
X_test = pd.DataFrame(missing_value_imputer.transform(X_test), columns = X_test.columns)

t1 = perf_counter()
print("Time Taken in minutes = ", np.round((t1 - t0)/60,2))

X_train.isnull().sum().sum()

X_test.isnull().sum().sum()

## We have successfully imputed missing values in the data

### 4. Data standardization

standard_scaler = StandardScaler()

# fit the standard scaler transformer on the train data
standard_scaler.fit(X_train)

X_train

## transform training and test feature     
transformed_std_scaled_train_df = pd.DataFrame(standard_scaler.transform(X_train))
transformed_std_scaled_train_df.columns = list(X_train.columns)

transformed_std_scaled_test_df = pd.DataFrame(standard_scaler.transform(X_test))
transformed_std_scaled_test_df.columns = list(X_test.columns)


### 5. Saving training and test data after imputation

X_train_df = transformed_std_scaled_train_df.copy()
X_test_df = transformed_std_scaled_test_df.copy()

X_train_df[target_variable] = y_train
X_test_df[target_variable] = y_test

X_train_df.to_csv("..\\1_Modeling_Data_After_Transformations_TestTrainSplits\\Training_data_std.csv", index  = False)
X_test_df.to_csv("..\\1_Modeling_Data_After_Transformations_TestTrainSplits\\Testing_data_std.csv", index  = False)
