==> Credit Default Prediction/2_Code_Jupyter_NoteBooks/04_HyperParameterOptimization_&_Saving_Best_Model.ipynb <==
### NoteBook Contents :
        > 1. Loading the Standard Scaled training data
        > 2. HyperParameter Tuning using Randomizaed Grid Search 5 Fold Cross Validation
        > 3. Training Best Classifier on the whole Trainig Data
        > 4. Saving the trained Model into the subdirectory

#### Importing Essential Data handling libraries

import pandas as pd
import numpy as np

import matplotlib as mpl
import matplotlib.pyplot as plt
import seaborn as sns

import warnings
warnings.filterwarnings("ignore")

%matplotlib inline

import time
from math import floor
from time import perf_counter
from tqdm import tqdm

pd.set_option('display.max_columns', 100)
pd.set_option('display.float_format', lambda x: '%.3f' % x)

from IPython.core.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

from sklearn.model_selection import train_test_split
from sklearn.model_selection import KFold
from sklearn.model_selection import cross_val_score
from sklearn import model_selection

from sklearn.metrics import f1_score

from sklearn.ensemble import RandomForestClassifier
import joblib

### 1. Loading the Standard Scaled training data

train_df = pd.read_csv("..\\1_Modeling_Data_After_Transformations_TestTrainSplits\\Training_data_std.csv") ## looking for dataset up one directory

train_df.head()

### 2. HyperParameter Tuning using Randomizaed Search Cross Validation

target = "default payment next month"

X = train_df.drop(target, axis = 1).values
y = train_df[target].values

##### Defining HyperParameter grid space and Fit classifier with 5 fold cross validation

# Number of trees in random forest
n_estimators = [250, 300, 400, 500]

# Maximum number of levels in tree
max_depth = [10, 15 ,20, 25, 30]

# Minimum number of samples required to split a node
min_samples_split = [2, 5, 10]

# Minimum number of samples required at each leaf node
min_samples_leaf = [1, 2, 4]

# Method of selecting samples for training each tree
bootstrap = [True, False]

param_grid = {'n_estimators': n_estimators,
               'max_depth': max_depth,
               'min_samples_split': min_samples_split,
               'min_samples_leaf': min_samples_leaf,
               'bootstrap': bootstrap}

param_grid

classifier = RandomForestClassifier(n_jobs = -1) # utilizing all the cores

model = model_selection.RandomizedSearchCV(estimator = classifier,
                                           param_distributions = param_grid,
                                           n_iter = 10,
                                           scoring = "f1", 
                                           verbose = 1,
                                           cv = 3,
                                           n_jobs = -1)

model.fit(X,y)

print(model.best_score_)
print(model.best_estimator_)

### 3. Training Best Classifier on the whole Trainig Data

final_model = model.best_estimator_

final_model.fit(X,y)

### 4. Saving the trained Model into the subdirectory

# save the model to disk
model_name = 'Random_Forest_Model_standard_scaled_data_v1.sav'
joblib.dump(final_model, "..\\3_Saved_Models"+"\\"+model_name)
