==> Credit Default Prediction/2_Code_Jupyter_NoteBooks/05_Model_Performance_on_Test_Data.ipynb <==
### NoteBook Contents :
        > 1. Loading the Preprocessed, transformed & Scaled Test Data
        > 2. Loading saved Model
        > 3. Prediction on test set

#### Importing Essential Data handling libraries

import pandas as pd
import numpy as np

import matplotlib as mpl
import matplotlib.pyplot as plt
import seaborn as sns

import warnings
warnings.filterwarnings("ignore")

%matplotlib inline

import time
from math import floor
from time import perf_counter
from tqdm import tqdm

pd.set_option('display.max_columns', 100)
pd.set_option('display.float_format', lambda x: '%.3f' % x)

from IPython.core.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

from sklearn.metrics import confusion_matrix
from sklearn.metrics import classification_report

from sklearn.ensemble import RandomForestClassifier
import joblib

### 1. Loading the Preprocessed, transformed & Scaled Test Data

test_df = pd.read_csv("..\\1_Modeling_Data_After_Transformations_TestTrainSplits\\Testing_data_std.csv") ## looking for dataset up one directory

test_df.head()

target = "default payment next month"

X_test = test_df.drop(target, axis = 1).values
y_test = test_df[target].values

###  2. Loading saved Model

model_path = '..\\3_Saved_Models\\Random_Forest_Model_standard_scaled_data_v1.sav'
model = joblib.load(model_path)

### 3. Prediction on test set and compute Accuracy score

y_pred = model.predict(X_test)

print(classification_report(y_test, y_pred))

print(confusion_matrix(y_test, y_pred))
